apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
tenacity==9.0.0
numpy==2.1.2
soundfile==0.12.1
